Rename 'IndividualCustomerPage' to 'CustomerViewPage'
