+++ src/features/customerView/CustomerViewPage.tsx
@@ -1,17 +1,17 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 
-import styles from './individualCustomerPage.module.scss';
+import styles from './customerViewPage.module.scss';
 import PageTitle from '../../common/pageTitle/PageTitle';
 
-const IndividualCustomerPage: React.SFC = ({ children }) => {
+const CustomerViewPage: React.SFC = ({ children }) => {
   const { t } = useTranslation();
   return (
-    <div className={styles.individualCustomerPage}>
+    <div className={styles.customerViewPage}>
       <PageTitle title={t('individualCustomer.title')} />
       <div className={styles.grid}>{children}</div>
     </div>
   );
 };
 
-export default IndividualCustomerPage;
+export default CustomerViewPage;
