+++ src/features/customerView/customerViewPage.module.scss
@@ -1,6 +1,6 @@
 @import 'spacings';
 
-.individualCustomerPage {
+.customerViewPage {
   min-height: 100%;
   padding: units(2);
 }
